Restart Geb VNC Container for each Test in a Spec to get correct recordings (#14995)
* test: proof of bug

test: proof of bug

* chore: boolean toggle

* fix using reflection

* test: improve clarity of test

And also, do not use an external url in the test.

* fix: rename property for clarity

* feat: default to restarting the recording container

* chore: cleanup

* fix: remove redundant system property setting

The system property
`grails.geb.recording.restartRecordingContainerPerTest`
was explicitly set to `true` for tests, but this is now
the default behavior, so the explicit setting is no longer
needed.

* test(geb): clean up old recording directories

Delete all but the two most recent recording directories
as part of the build to avoid unnecessary disk usage.

* Restore Slf4j logging to WebDriverContainerHolder

Restore Slf4j logging lost in the merge

---------

Co-authored-by: Mattias Reichel <mattias.reichel@gmail.com>
Co-authored-by: James Fredley <jamesfredley@users.noreply.github.com>